requirements.txt
Pillow
pytest
aiohttp
aiofiles
requests
asyncio_timeout